refactor(storage): cohesive barrier boundary, and drop the extra contract module
Three findings, all valid.

**The contract module is gone.** I created it to break an import cycle, and
checking rather than assuming: there was no cycle to break. The coordinator
imports nothing from the scheduler, and the scheduler already imports the
coordinator, so the shared declarations live in the coordinator module and the
third file was pure indirection.

**The host is a boundary, not a bag of counters.** Eight fine-grained accessors
become `quiescence(storeId)` — the actual question, asked once — plus
`blockerDiagnostics()`, and the split is load-bearing rather than cosmetic:
`pump()` runs on every completion, while `countGenerationsInflight()` is
O(stores). Collapsing everything into one call would have put an O(stores) scan
on the hot path, so the diagnostics stay separate and are read only when the
bound expires.

`isReady` now reads as the question it asks — no untagged work, none tagged for
this store — instead of two unrelated accessor calls.

**Orphaned comment removed.** The moved constant carries its own documentation
in the coordinator; the stub left behind pointed at nothing.

Re-verified discrimination after the reshape rather than trusting the earlier
run: the boundary the mutant targets changed shape, so zeroing
`taggedInflightForStore` inside the new `quiescence` callback still reddens
"waits for BOTH untagged and same-store tagged work before starting".

Storage lane 31 files / 528 tests, typecheck clean.
